fix : 카테고리, 검색 페이지 공통 부분 item_list 컴포넌트로 추출 #33
