+++ app/(main)/search/page.tsx
@@ -9,6 +9,12 @@ export default async function SearchPage({ searchParams }: { searchParams: { pag
   const currentPage = Number(searchParams.page) || 1;
   const itemsPerPage = 20;
 
+  // 페이지네이션 유효성 검사
+  const totalPages = Math.ceil(items.length / itemsPerPage);
+  if (currentPage < 1 || currentPage > totalPages) {
+    return { notFound: true };
+  }
+
   // 현재 페이지에 해당하는 아이템들만 가져오기
   const startIndex = (currentPage - 1) * itemsPerPage;
   const endIndex = startIndex + itemsPerPage;
