+++ components/common/item_list.module.css
@@ -1,0 +1,26 @@
+.item_list__page {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    padding: 40px 20px;
+    width: 100%;
+    max-width: 1080px;
+    margin: 0 auto;
+}
+
+.item_list__title {
+    text-align: left;
+    margin: 30px 0 50px 0;
+}
+
+.item_list__item_list {
+    display: grid;
+    grid-template-columns: repeat(4, 1fr);
+    gap: 120px 20px;
+    width: 100%;
+    justify-content: center;
+}
+
+.item_list__pagination {
+    margin-top: 120px;
+}
